nostradamus/requirements.txt
Django==3.0.3
djangorestframework==3.11.0
jira==2.0.0
pandas==1.0.3
imbalanced-learn==0.4.2
pytest==5.3.5
scikit-learn==0.20.3
nltk==3.4.5
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
pre-commit==2.2.0
redis==3.4.1
mongoengine==0.19.1
gunicorn==20.0.4
celery==4.4.2
daphne==2.5.0
channels-redis==2.4.2
channels==2.4.0
pystemmer==2.0.0.1
uvicorn==0.11.5
flower==0.9.4